Refactored login code
